Autoloader working properly with createPhar.php

+++ daemon/source/createPhar.php
@@ -6,6 +6,11 @@ $archiveBaseName = "vagrant_sync";
 $archiveName = "{$archiveBaseName}.phar";
 $archivePath = $buildRoot . DIRECTORY_SEPARATOR . $archiveName;
 
+if (ini_get('phar.readonly')) {
+	echo 'phar.readonly must be set to false before running this script!';
+	exit(2);
+}
+$filesToAdd = array();
 if (file_exists($archivePath)) {
 	if (!unlink($archivePath)) {
 		throw new Exception("Unable to delete file $archivePath");
@@ -16,7 +21,9 @@ $phar = new Phar(
 	(FilesystemIterator::CURRENT_AS_FILEINFO | FilesystemIterator::KEY_AS_FILENAME),
 	$archiveName
 );
-$phar["autoloader.php"] = file_get_contents($srcRoot . DIRECTORY_SEPARATOR . "autoloader.php");
+$phar->compressFiles(Phar::BZ2);
+$phar->setSignatureAlgorithm(Phar::SHA512);
 $phar->setStub(file_get_contents($srcRoot . DIRECTORY_SEPARATOR . "stub.php"));
+$phar->buildFromDirectory($srcRoot, ',/(autoloader|namespace/.*)\.php$,');
 
 chmod($archivePath, 0755);
